requirements.txt
airflow-code-editor==5.2.0
alembic==1.7.5
anyio==3.4.0
apache-airflow==2.2.3
apache-airflow-providers-ftp==2.0.1
apache-airflow-providers-http==2.0.1
apache-airflow-providers-imap==2.0.1
apache-airflow-providers-snowflake==2.3.0
apache-airflow-providers-sqlite==2.0.1
apispec==3.3.2
appdirs==1.4.4
argcomplete==1.12.3
asn1crypto==1.5.1
attrs==20.3.0
azure-common==1.1.28
azure-core==1.23.1
azure-storage-blob==12.11.0
Babel==2.9.1
beautifulsoup4==4.11.1
blinker==1.4
boto3==1.21.37
botocore==1.24.37
bs4==0.0.1
cachelib==0.6.0
cattrs==1.5.0
certifi==2020.12.5
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.9
click==8.0.3
clickclick==20.10.2
colorama==0.4.4
colorlog==4.8.0
commonmark==0.9.1
croniter==1.0.15
cryptography==3.4.8
cvxpy==1.2.0
cycler==0.11.0
defusedxml==0.7.1
dill==0.3.1.1
dnspython==2.1.0
docutils==0.16
ecos==2.0.10
email-validator==1.1.3
exchange-calendars==3.6.1
Flask==1.1.2
Flask-AppBuilder==3.4.1
Flask-Babel==2.0.0
Flask-Caching==1.10.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
fonttools==4.32.0
graphviz==0.19.1
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
httpcore==0.13.7
httpx==0.19.0
idna==2.10
importlib-metadata==4.8.2
importlib-resources==5.4.0
inflection==0.5.1
iso8601==1.0.2
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==3.0.3
jmespath==1.0.0
joblib==1.1.0
jsonschema==3.2.0
kiwisolver==1.4.2
korean-lunar-calendar==0.2.1
lazy-object-proxy==1.4.3
lockfile==0.12.2
lxml==4.8.0
Mako==1.1.6
Markdown==3.3.6
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.5.1
msrest==0.6.21
multitasking==0.0.10
nltk==3.7
numpy==1.22.3
oauthlib==3.2.0
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.1
oscrypto==1.3.0
osqp==0.6.2.post5
packaging==21.3
pandas==1.4.1
pandas-datareader==0.10.0
pendulum==2.1.2
Pillow==9.1.0
prison==0.2.1
psutil==5.8.0
pycparser==2.21
pycryptodomex==3.14.1
Pygments==2.10.0
PyJWT==1.7.1
pyluach==1.4.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.18.0
python-daemon==2.3.0
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2021.3
pytzdata==2020.1
PyYAML==5.4.1
qdldl==0.1.5.post2
regex==2022.3.15
requests==2.26.0
requests-cache==0.9.3
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==10.16.0
s3transfer==0.5.2
scipy==1.8.0
scs==3.2.0
seaborn==0.11.2
setproctitle==1.2.2
six==1.16.0
sniffio==1.2.0
snowflake-connector-python==2.4.3
snowflake-sqlalchemy==1.3.3
SQLAlchemy==1.4.35
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.37.9
swagger-ui-bundle==0.0.9
tabulate==0.8.9
tenacity==8.0.1
termcolor==1.1.0
text-unidecode==1.3
toolz==0.11.2
tqdm==4.64.0
typing-extensions==4.1.1
ujson==5.2.0
unicodecsv==0.14.1
url-normalize==1.4.3
urllib3==1.26.7
Werkzeug==1.0.1
WTForms==2.3.3
yfinance==0.1.70
zipp==3.6.0
